requirements copy.txt
click==7.1.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.3
joblib==0.17.0
numpy==1.19.4
opencv-contrib-python==4.2.0.34
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
threadpoolctl==2.1.0
face-recognition==1.3.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
Pillow==10.0.0
streamlit==1.24.0
pathlib==1.0.1
psutil==5.9.5
numpy>=1.19.3
dlib>=19.22.0